Setup Front End Unit Class
This class is a replica of the WPCW_UnitFrontend in WP-Courseware. I
wasn’t able to extend the class and I was not able to hook into any
actions or filters to achieve our desired look and feel for the
“Completed” boxes, Navigation Links, and AJAX functions that are built
into WP Courseware.

This class lets us customize the functionality while maintaining the
integrity of the core plugin.

*Note - the “the_content” filter that WP-Courseware adds in its main
plugin file was removed from the core plugin. There was no other way to
do it. Be mindful of that whenever updating WP-Courseware