Pass DB data over to the main.js file
Using wp_localize_script() we are passing data over to the main.js
file.

I setup a sample of how the data is passed into the script and how to
access that data. You can also pass the entire global object into the
data but the lack of underscores in the database key name was causing
me some issues when displaying the data in the alert box.

Passing in only what we need should be sufficient since we only need a
few things.

In the i4-class-admin-menu.php file is where you’ll find the name of
any of the i-4Web LMS settings. i.e. i4-lms-course-force-viewing